requirements.txt

# AI and Vector Database
anthropic>=0.25.0
astrapy>=1.0.0
voyageai>=0.2.0

# Keep existing requirements
# Pure AI Trading Agent Requirements
# Updated for Claude Sonnet 4 + AstraDB + Tool-based Architecture

# AI and Language Models
anthropic>=0.25.0
langchain-anthropic==0.3.0

# Vector Database
astrapy>=1.0.0

# Embedding Services (Optional - for better embeddings)
voyageai>=0.2.0

# Existing Core Dependencies
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrdict==2.0.1
attrs==25.3.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.3.0
bip-utils==2.9.3
blinker==1.9.0
boto3==1.37.37
botocore==1.37.37
build==1.2.2.post1
cachetools==5.5.2
cbor2==5.6.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.0
coincurve==21.0.0
coloredlogs==15.0.1
construct==2.10.68
construct-typing==0.5.6
crcmod==1.7
cytoolz==0.12.3
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
ecdsa==0.19.1
ed25519-blake2b==1.4.1
eth-abi==1.3.0
eth-account==0.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
faiss-cpu==1.11.0
fastapi==0.115.9
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonalias==0.1.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.42
lru-dict==1.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.39.0
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.22.0

# Keep OpenAI for fallback embeddings only
openai==1.77.0

opentelemetry-api==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-instrumentation-asgi==0.54b0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opentelemetry-util-http==0.54b0
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parsimonious==0.8.1
pillow==11.2.1
plotly==6.0.1
posthog==4.0.1
propcache==0.3.1
protobuf==5.29.4
py-sr25519-bindings==0.2.2
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyNaCl==1.5.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rlp==2.0.1
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.11.5
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
SQLAlchemy==2.0.40
starlette==0.45.3
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
web3==4.2.1
websocket-client==1.8.0
websockets==15.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0